+++ src/extensions/index.ts
@@ -1,3 +1,4 @@
 export * from './assetsPathPrefix/assetsPathPrefix.extension';
 export * from './gfmHeadingId/gfmHeadingId.extension';
 export * from './includeMarkdown/includeMarkdown.extension';
+export * from './tableMarkdownIndex/tableMarkdownIndex.extension';